// apps/dispatch-service/src/enums/trip-status.enum.ts
export enum TripStatus {
  CREATED = 'CREATED',
  DISPATCHED = 'DISPATCHED',
  EN_ROUTE_SCENE = 'EN_ROUTE_SCENE',
  AT_SCENE = 'AT_SCENE',
  PATIENT_LOADED = 'PATIENT_LOADED',
  EN_ROUTE_HOSPITAL = 'EN_ROUTE_HOSPITAL',
  AT_HOSPITAL = 'AT_HOSPITAL',
  HANDOFF_COMPLETE = 'HANDOFF_COMPLETE',
  CANCELLED = 'CANCELLED',
  BREAKDOWN = 'BREAKDOWN',
}